backend/requirements.txt
amqp==5.3.1
argcomplete==3.6.3
asgiref==3.10.0
attrs==23.2.0
autobahn==25.10.2
Automat==22.10.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
billiard==4.2.3
blinker==1.7.0
boto3==1.34.46
botocore==1.34.46
celery==5.5.3
certifi==2023.11.17
channels==4.3.1
channels_redis==4.3.0
chardet==5.2.0
click==8.1.6
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloud-init==25.1.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
coverage==7.11.3
cryptography==41.0.7
daphne==4.2.1
dbus-python==1.3.2
distro==1.9.0
distro-info==1.7+build1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==23.0.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
iniconfig==2.3.0
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
kombu==5.5.4
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mercurial==6.7.2
msgpack==1.1.2
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.2
packaging==24.0
pexpect==4.9.0
pipx==1.8.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyparted==3.12.0
pyrsistent==0.20.0
pyserial==3.5
pytest==9.0.1
pytest-cov==7.0.0
pytest-django==4.11.1
python-apt==2.7.7+ubuntu5
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
redis==7.0.1
requests==2.31.0
rich==13.7.1
s3transfer==0.10.1
service-identity==24.1.0
setuptools==68.1.2
six==1.16.0
sos==4.8.2
sqlparse==0.5.3
ssh-import-id==5.11
systemd-python==235
Twisted==24.3.0
txaio==25.9.2
typing_extensions==4.10.0
tzdata==2025.2
ubuntu-pro-client==8001
ufw==0.36.2
uritemplate==4.2.0
urllib3==2.0.7
userpath==1.9.2
vine==5.1.0
wadllib==1.3.6
WALinuxAgent==2.11.1.4
wcwidth==0.2.14
wheel==0.42.0
zope.interface==6.1
zstandard==0.22.0
